Major refactor of the query code
