+++ data_subscriber/query.py
@@ -14,294 +14,144 @@ import json
 import dateutil.parser
 from hysds_commons.job_utils import submit_mozart_job
 from more_itertools import chunked, first
-
 from data_subscriber.cmr import COLLECTION_TO_PRODUCT_TYPE_MAP, async_query_cmr, CMR_COLLECTION_TO_PROVIDER_TYPE_MAP
 from data_subscriber.hls.hls_catalog import HLSProductCatalog
-from data_subscriber.hls_spatial.hls_spatial_catalog_connection import get_hls_spatial_catalog_connection
-from data_subscriber.rtc import mgrs_bursts_collection_db_client as mbc_client, evaluator
-from data_subscriber.rtc.rtc_download_job_submitter import submit_rtc_download_job_submissions_tasks
-from data_subscriber.slc_spatial.slc_spatial_catalog_connection import get_slc_spatial_catalog_connection
 from data_subscriber.url import form_batch_id, form_batch_id_cslc, _slc_url_to_chunk_id
-from data_subscriber.cslc_utils import localize_disp_frame_burst_json, expand_clsc_frames, build_cslc_native_ids
+from data_subscriber.geojson_utils import localize_geojsons, localize_include_exclude, filter_granules_by_regions
+from data_subscriber.rtc.rtc_download_job_submitter import submit_rtc_download_job_submissions_tasks
 from geo.geo_util import does_bbox_intersect_north_america, does_bbox_intersect_region, _NORTH_AMERICA
-from rtc_utils import rtc_granule_regex
 from util.conf_util import SettingsConf
 
 logger = logging.getLogger(__name__)
 
 DateTimeRange = namedtuple("DateTimeRange", ["start_date", "end_date"])
 
-def catalog_granules(granule, args, job_id, es_conn, download_batch_id, query_dt, affected_mgrs_set_id_acquisition_ts_cycle_indexes):
-    granule_id = granule.get("granule_id")
-    revision_id = granule.get("revision_id")
-
-    additional_fields = {}
-    if download_batch_id is not None:
-        additional_fields["download_batch_id"] = download_batch_id
-    additional_fields["revision_id"] = revision_id
-    additional_fields["processing_mode"] = args.proc_mode
-
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
-        additional_fields["instrument"] = "S1A" if "S1A" in granule_id else "S1B"
-
-        match_product_id = re.match(rtc_granule_regex, granule_id)
-        acquisition_dts = match_product_id.group("acquisition_ts")
-        burst_id = match_product_id.group("burst_id")
-
-        mgrs = mbc_client.cached_load_mgrs_burst_db(filter_land=True)
-        mgrs_burst_set_ids = mbc_client.burst_id_to_mgrs_set_ids(mgrs, mbc_client.product_burst_id_to_mapping_burst_id(
-            burst_id))
-        additional_fields["mgrs_set_ids"] = mgrs_burst_set_ids
-
-        # RTC: Calculating the Collection Cycle Index (Part 1):
-        #  required constants
-        MISSION_EPOCH_S1A = dateutil.parser.isoparse("20190101T000000Z")  # set approximate mission start date
-        MISSION_EPOCH_S1B = MISSION_EPOCH_S1A + timedelta(days=6)  # S1B is offset by 6 days
-        MAX_BURST_IDENTIFICATION_NUMBER = 375887  # gleamed from MGRS burst collection database
-        ACQUISITION_CYCLE_DURATION_SECS = timedelta(days=12).total_seconds()
-
-        # RTC: Calculating the Collection Cycle Index (Part 2):
-        #  RTC products can be indexed into their respective elapsed collection cycle since mission start/epoch.
-        #  The cycle restarts periodically with some miniscule drift over time and the life of the mission.
-        burst_identification_number = int(burst_id.split(sep="-")[1])
-        instrument_epoch = MISSION_EPOCH_S1A if "S1A" in granule_id else MISSION_EPOCH_S1B
-        seconds_after_mission_epoch = (dateutil.parser.isoparse(acquisition_dts) - instrument_epoch).total_seconds()
-        acquisition_index = (
-                                    seconds_after_mission_epoch - (ACQUISITION_CYCLE_DURATION_SECS * (
-                                        burst_identification_number / MAX_BURST_IDENTIFICATION_NUMBER))
-                            ) / ACQUISITION_CYCLE_DURATION_SECS
-        acquisition_cycle = round(acquisition_index)
-        additional_fields["acquisition_cycle"] = acquisition_cycle
-
-        update_additional_fields_mgrs_set_id_acquisition_ts_cycle_indexes(acquisition_cycle, acquisition_index,
-                                                                          additional_fields, mgrs_burst_set_ids)
-        update_affected_mgrs_set_ids(acquisition_cycle, acquisition_index,
-                                     affected_mgrs_set_id_acquisition_ts_cycle_indexes, mgrs_burst_set_ids)
-
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "SLC":
-        if does_bbox_intersect_north_america(granule["bounding_box"]):
-            additional_fields["intersects_north_america"] = True
-    elif COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "CSLC":
-        pass
-    else:
-        pass
-
-    update_url_index(
-        es_conn,
-        granule.get("filtered_urls"),
-        granule_id,
-        job_id,
-        query_dt,
-        temporal_extent_beginning_dt=dateutil.parser.isoparse(granule["temporal_extent_beginning_datetime"]),
-        revision_date_dt=dateutil.parser.isoparse(granule["revision_date"]),
-        **additional_fields
-    )
-
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "HLS":
-        spatial_catalog_conn = get_hls_spatial_catalog_connection(logger)
-        update_granule_index(spatial_catalog_conn, granule)
-    elif COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "SLC":
-        spatial_catalog_conn = get_slc_spatial_catalog_connection(logger)
-        update_granule_index(spatial_catalog_conn, granule)
-    elif COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
-        pass
-    elif COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "CSLC":
-        pass
-    else:
-        pass
-
 async def run_query(args, token, es_conn: HLSProductCatalog, cmr, job_id, settings):
-    query_dt = datetime.now()
-    now = datetime.utcnow()
-    query_timerange: DateTimeRange = get_query_timerange(args, now)
-    download_batch_id = None
+    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "HLS":
+        from data_subscriber.hls.hls_query import HlsCmrQuery
+        cmr_query = HlsCmrQuery(args, token, es_conn, cmr, job_id, settings)
+    elif COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "SLC":
+        from data_subscriber.slc.slc_query import SlcCmrQuery
+        cmr_query = SlcCmrQuery(args, token, es_conn, cmr, job_id, settings)
+    elif COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
+        from data_subscriber.rtc.rtc_query import RtcCmrQuery
+        cmr_query = RtcCmrQuery(args, token, es_conn, cmr, job_id, settings)
+    elif COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "CSLC":
+        from data_subscriber.cslc.cslc_query import CslcCmrQuery
+        cmr_query = CslcCmrQuery(args, token, es_conn, cmr, job_id, settings)
 
-    # If we are querying CSLC data we need to modify parameters going into cmr query
-    #TODO: put this in a loop and query one frame at a time
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "CSLC" and args.frame_range is not None:
-        disp_burst_map, metadata, version = localize_disp_frame_burst_json()
+    result = await cmr_query.run_query(args, token, es_conn, cmr, job_id, settings)
+    return result
 
-        #TODO: If we process more than one frame in a single query, we need to restructure this.
-        # Use underscore instead of other special characters and lower case so that it can be used in ES term search
-        download_batch_id = args.start_date+"_"+args.end_date+"_"+args.frame_range.split(",")[0]
-        download_batch_id = download_batch_id.replace("-", "_").replace(":", "_").lower()
+class CmrQuery:
+    def __init__(self, args, token, es_conn, cmr, job_id, settings):
+        self.args = args
+        self.token = token
+        self.es_conn = es_conn
+        self.cmr = cmr
+        self.job_id = job_id
+        self.settings = settings
+        self.affected_mgrs_set_id_acquisition_ts_cycle_indexes = set()
 
-        if expand_clsc_frames(args, disp_burst_map) == False:
-            logging.info("No valid frames were found.")
+    async def run_query(self, args, token, es_conn: HLSProductCatalog, cmr, job_id, settings):
+        query_dt = datetime.now()
+        now = datetime.utcnow()
+        query_timerange: DateTimeRange = get_query_timerange(args, now)
+
+        logger.info("CMR query STARTED")
+        granules, download_batch_id = await self.query_cmr(args, token, cmr, settings, query_timerange, now)
+        logger.info("CMR query FINISHED")
+
+        if args.smoke_run:
+            logger.info(f"{args.smoke_run=}. Restricting to 1 granule(s).")
+            granules = granules[:1]
+
+        # If processing mode is historical, apply include/exclude-region filtering
+        if args.proc_mode == "historical":
+            logging.info(f"Processing mode is historical so applying include and exclude regions...")
+
+            # Fetch all necessary geojson files from S3
+            localize_include_exclude(args)
+            granules[:] = filter_granules_by_regions(granules, args.include_regions, args.exclude_regions)
+
+        logger.info("catalogue-ing STARTED")
+        self.catalog_granules(granules, args, job_id, es_conn, download_batch_id, query_dt)
+        logger.info("catalogue-ing FINISHED")
+
+        batch_id_to_products_map = await self.refresh_index(args, es_conn)
+
+        if args.subparser_name == "full":
+            logger.info(
+                f"{args.subparser_name=}. Skipping download job submission. Download will be performed directly.")
+            if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
+                args.provider = CMR_COLLECTION_TO_PROVIDER_TYPE_MAP[args.collection]
+                args.batch_ids = self.affected_mgrs_set_id_acquisition_ts_cycle_indexes
+            return
+        if args.no_schedule_download:
+            logger.info(f"{args.no_schedule_download=}. Forcefully skipping download job submission.")
+            return
+        if not args.chunk_size:
+            logger.info(f"{args.chunk_size=}. Insufficient chunk size. Skipping download job submission.")
             return
 
-    logger.info("CMR query STARTED")
-    granules = await async_query_cmr(args, token, cmr, settings, query_timerange, now)
-    logger.info("CMR query FINISHED")
-
-    if args.smoke_run:
-        logger.info(f"{args.smoke_run=}. Restricting to 1 granule(s).")
-        granules = granules[:1]
-
-    # If we are processing ASF collection, we're gonna need the north america geojson
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "SLC":
-        localize_geojsons([_NORTH_AMERICA])
-
-    # If processing mode is historical, apply include/exclude-region filtering
-    if args.proc_mode == "historical":
-        logging.info(f"Processing mode is historical so applying include and exclude regions...")
-
-        # Fetch all necessary geojson files from S3
-        localize_include_exclude(args)
-        granules[:] = filter_granules_by_regions(granules, args.include_regions, args.exclude_regions)
-
-    logger.info("catalogue-ing STARTED")
-
-    affected_mgrs_set_id_acquisition_ts_cycle_indexes = None
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
-        affected_mgrs_set_id_acquisition_ts_cycle_indexes = set()
-        granules[:] = filter_granules_rtc(granules, args)
-
-    for granule in granules:
-        catalog_granules(granule, args, job_id, es_conn, download_batch_id, query_dt, affected_mgrs_set_id_acquisition_ts_cycle_indexes)
-
-    logger.info("catalogue-ing FINISHED")
-
-    succeeded = []
-    failed = []
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
-        logger.info("performing index refresh")
-        es_conn.refresh()
-        logger.info("performed index refresh")
-
-        logger.info("evaluating available burst sets")
-        logger.info(f"{affected_mgrs_set_id_acquisition_ts_cycle_indexes=}")
-        fully_covered_mgrs_sets, target_covered_mgrs_sets, incomplete_mgrs_sets = await evaluator.main(
-            mgrs_set_id_acquisition_ts_cycle_indexes=affected_mgrs_set_id_acquisition_ts_cycle_indexes,
-            coverage_target=settings["DSWX_S1_COVERAGE_TARGET"]
-        )
-
-        processable_mgrs_sets = {**incomplete_mgrs_sets, **fully_covered_mgrs_sets}
-
-        # convert to "batch_id" mapping
-        batch_id_to_products_map = defaultdict(set)
-        for mgrs_set_id, product_burst_sets in processable_mgrs_sets.items():
-            for product_burstset in product_burst_sets:
-                rtc_granule_id_to_product_docs_map = first(product_burstset)
-                first_product_doc_list = first(rtc_granule_id_to_product_docs_map.values())
-                first_product_doc = first(first_product_doc_list)
-                acquisition_cycle = first_product_doc["acquisition_cycle"]
-                batch_id = "{}${}".format(mgrs_set_id, acquisition_cycle)
-                batch_id_to_products_map[batch_id] = product_burstset
-                if args.smoke_run:
-                    logger.info(f"{args.smoke_run=}. Not processing more burst_sets.")
-                    break
-            if args.smoke_run:
-                logger.info(f"{args.smoke_run=}. Not processing more sets of burst_sets.")
-                break
-
-    if args.subparser_name == "full":
-        logger.info(f"{args.subparser_name=}. Skipping download job submission. Download will be performed directly.")
         if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
-            args.provider = CMR_COLLECTION_TO_PROVIDER_TYPE_MAP[args.collection]
-            args.batch_ids = affected_mgrs_set_id_acquisition_ts_cycle_indexes
-        return
-    if args.no_schedule_download:
-        logger.info(f"{args.no_schedule_download=}. Forcefully skipping download job submission.")
-        return
-    if not args.chunk_size:
-        logger.info(f"{args.chunk_size=}. Insufficient chunk size. Skipping download job submission.")
-        return
+            job_submission_tasks = submit_rtc_download_job_submissions_tasks(batch_id_to_products_map.keys(), args)
+        else:
+            job_submission_tasks = download_job_submission_handler(args, granules, query_timerange, download_batch_id)
 
-    if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
-        job_submission_tasks = submit_rtc_download_job_submissions_tasks(batch_id_to_products_map.keys(), args)
-    else:
-        job_submission_tasks = download_job_submission_handler(args, granules, query_timerange, download_batch_id)
+        results = await asyncio.gather(*job_submission_tasks, return_exceptions=True)
+        logger.info(f"{len(results)=}")
+        logger.info(f"{results=}")
 
-    results = await asyncio.gather(*job_submission_tasks, return_exceptions=True)
-    logger.info(f"{len(results)=}")
-    logger.info(f"{results=}")
+        succeeded = [job_id for job_id in results if isinstance(job_id, str)]
+        failed = [e for e in results if isinstance(e, Exception)]
 
-    succeeded = [job_id for job_id in results if isinstance(job_id, str)]
-    failed = [e for e in results if isinstance(e, Exception)]
+        logger.info(f"{succeeded=}")
+        logger.info(f"{failed=}")
 
-    logger.info(f"{succeeded=}")
-    logger.info(f"{failed=}")
+        return {
+            "success": succeeded,
+            "fail": failed
+        }
 
-    return {
-        "success": succeeded,
-        "fail": failed
-    }
+    async def query_cmr(self, args, token, cmr, settings, timerange, now: datetime):
+        granules = await async_query_cmr(args, token, cmr, settings, timerange, now)
+        return granules, None
 
+    def prepare_additional_fields(self, granule, args, granule_id, download_batch_id):
 
-def update_affected_mgrs_set_ids(acquisition_cycle, acquisition_index, affected_mgrs_set_id_acquisition_ts_cycle_indexes, mgrs_burst_set_ids):
-    acquisition_index_floor = math.floor(acquisition_index)
-    acquisition_index_ceil = math.ceil(acquisition_index)
-    # construct filters for evaluation
-    if len(mgrs_burst_set_ids) == 1:
-        # ati = Acquisition Time Index
-        if acquisition_cycle == acquisition_index_floor:  # rounded down, closer to start of cycle
-            current_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
-            future_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle + 1)
+        additional_fields = {}
+        if download_batch_id is not None:
+            additional_fields["download_batch_id"] = download_batch_id
+        additional_fields["revision_id"] = granule.get("revision_id")
+        additional_fields["processing_mode"] = args.proc_mode
 
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(current_ati)
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(future_ati)
+        return additional_fields
 
-        if acquisition_cycle == acquisition_index_ceil:  # rounded up, closer to end of cycle
-            past_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle - 1)
-            current_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
+    def catalog_granules(self, granules, args, job_id, es_conn, download_batch_id, query_dt):
+        for granule in granules:
+            granule_id = granule.get("granule_id")
 
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(past_ati)
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(current_ati)
-    elif len(mgrs_burst_set_ids) == 2:
-        if acquisition_cycle == acquisition_index_floor:  # rounded down, closer to start of cycle
-            current_ati_a = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
-            current_ati_b = "{}${}".format(sorted(mgrs_burst_set_ids)[1], acquisition_cycle)
-            future_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[1], acquisition_cycle + 1)
+            additional_fields = self.prepare_additional_fields(granule, args, granule_id, download_batch_id)
 
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(current_ati_a)
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(current_ati_b)
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(future_ati)
-        if acquisition_cycle == acquisition_index_ceil:  # rounded up, closer to end of cycle
-            past_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle - 1)
-            current_ati_a = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
-            current_ati_b = "{}${}".format(sorted(mgrs_burst_set_ids)[1], acquisition_cycle)
+            update_url_index(
+                es_conn,
+                granule.get("filtered_urls"),
+                granule_id,
+                job_id,
+                query_dt,
+                temporal_extent_beginning_dt=dateutil.parser.isoparse(granule["temporal_extent_beginning_datetime"]),
+                revision_date_dt=dateutil.parser.isoparse(granule["revision_date"]),
+                **additional_fields
+            )
 
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(past_ati)
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(current_ati_a)
-            affected_mgrs_set_id_acquisition_ts_cycle_indexes.add(current_ati_b)
-    else:
-        raise AssertionError("Unexpected burst overlap")
+            self.update_granule_index(granule)
 
+    def update_granule_index(self, granule):
+        pass
 
-def update_additional_fields_mgrs_set_id_acquisition_ts_cycle_indexes(acquisition_cycle, acquisition_index, additional_fields, mgrs_burst_set_ids):
-    acquisition_index_floor = math.floor(acquisition_index)
-    acquisition_index_ceil = math.ceil(acquisition_index)
-    # construct filters for evaluation
-    if len(mgrs_burst_set_ids) == 1:
-        # ati = Acquisition Time Index
-        if acquisition_cycle == acquisition_index_floor:  # rounded down, closer to start of cycle
-            current_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
-            future_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle + 1)
-
-            additional_fields["mgrs_set_id_acquisition_ts_cycle_indexes"] = [current_ati]
-
-        if acquisition_cycle == acquisition_index_ceil:  # rounded up, closer to end of cycle
-            past_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle - 1)
-            current_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
-
-            additional_fields["mgrs_set_id_acquisition_ts_cycle_indexes"] = [current_ati]
-    elif len(mgrs_burst_set_ids) == 2:
-        if acquisition_cycle == acquisition_index_floor:  # rounded down, closer to start of cycle
-            current_ati_a = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
-            current_ati_b = "{}${}".format(sorted(mgrs_burst_set_ids)[1], acquisition_cycle)
-            future_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[1], acquisition_cycle + 1)
-
-            additional_fields["mgrs_set_id_acquisition_ts_cycle_indexes"] = [current_ati_a, current_ati_b]
-        if acquisition_cycle == acquisition_index_ceil:  # rounded up, closer to end of cycle
-            past_ati = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle - 1)
-            current_ati_a = "{}${}".format(sorted(mgrs_burst_set_ids)[0], acquisition_cycle)
-            current_ati_b = "{}${}".format(sorted(mgrs_burst_set_ids)[1], acquisition_cycle)
-
-            additional_fields["mgrs_set_id_acquisition_ts_cycle_indexes"] = [current_ati_a, current_ati_b]
-    else:
-        raise AssertionError("Unexpected burst overlap")
-
+    async def refresh_index(self, args, es_conn):
+        pass
 
 def download_job_submission_handler(args, granules, query_timerange, download_batch_id):
     batch_id_to_urls_map = defaultdict(set)
@@ -335,9 +185,10 @@ def download_job_submission_handler(args, granules, query_timerange, download_ba
         job_submission_tasks = submit_download_job_submissions_tasks(batch_id_to_urls_map, query_timerange, args)
     return job_submission_tasks
 
-
 def get_query_timerange(args, now: datetime, silent=False):
-    now_minus_minutes_dt = (now - timedelta(minutes=args.minutes)) if not args.native_id else dateutil.parser.isoparse("1900-01-01T00:00:00Z")
+    now_minus_minutes_dt = (
+                now - timedelta(minutes=args.minutes)) if not args.native_id else dateutil.parser.isoparse(
+        "1900-01-01T00:00:00Z")
 
     start_date = args.start_date if args.start_date else now_minus_minutes_dt.strftime("%Y-%m-%dT%H:%M:%SZ")
     end_date = args.end_date if args.end_date else now.strftime("%Y-%m-%dT%H:%M:%SZ")
@@ -346,7 +197,6 @@ def get_query_timerange(args, now: datetime, silent=False):
     if not silent:
         logger.info(f"{query_timerange=}")
     return query_timerange
-
 
 def submit_download_job_submissions_tasks(batch_id_to_urls_map, query_timerange, args):
     job_submission_tasks = []
@@ -484,35 +334,6 @@ def update_url_index(
     for filename, filename_urls in filename_to_urls_map.items():
         es_conn.process_url(filename_urls, granule_id, job_id, query_dt, temporal_extent_beginning_dt, revision_date_dt, *args, **kwargs)
 
-
-def update_granule_index(es_spatial_conn, granule, *args, **kwargs):
-    es_spatial_conn.process_granule(granule, *args, **kwargs)
-
-
-def localize_include_exclude(args):
-
-    geojsons = []
-
-    if args.include_regions is not None:
-        geojsons.extend(args.include_regions.split(","))
-
-    if args.exclude_regions is not None:
-        geojsons.extend(args.exclude_regions.split(","))
-
-    localize_geojsons(geojsons)
-
-
-def localize_geojsons(geojsons):
-    settings = SettingsConf().cfg
-    bucket = settings["GEOJSON_BUCKET"]
-
-    try:
-        for geojson in geojsons:
-            key = geojson.strip() + ".geojson"
-            # output_filepath = os.path.join(working_dir, key)
-            download_from_s3(bucket, key, key)
-    except Exception as e:
-        raise Exception("Exception while fetching geojson file: %s. " % key + str(e))
 def process_frame_burst_db():
     settings = SettingsConf().cfg
     bucket = settings["GEOJSON_BUCKET"]
@@ -524,68 +345,3 @@ def process_frame_burst_db():
             download_from_s3(bucket, key, key)
     except Exception as e:
         raise Exception("Exception while fetching geojson file: %s. " % key + str(e))
-
-def does_granule_intersect_regions(granule, intersect_regions):
-    regions = intersect_regions.split(',')
-    for region in regions:
-        region = region.strip()
-        if does_bbox_intersect_region(granule["bounding_box"], region):
-            return True, region
-
-    return False, None
-
-
-def filter_granules_by_regions(granules, include_regions, exclude_regions):
-    '''Filters granules based on include and exclude regions lists'''
-    filtered = []
-
-    for granule in granules:
-
-        # Skip this granule if it's not in the include list
-        if include_regions is not None:
-            (result, region) = does_granule_intersect_regions(granule, include_regions)
-            if result is False:
-                logging.info(
-                    f"The following granule does not intersect with any include regions. Skipping processing %s"
-                    % granule.get("granule_id"))
-                continue
-
-        # Skip this granule if it's in the exclude list
-        if exclude_regions is not None:
-            (result, region) = does_granule_intersect_regions(granule, exclude_regions)
-            if result is True:
-                logging.info(f"The following granule intersects with the exclude region %s. Skipping processing %s"
-                             % (region, granule.get("granule_id")))
-                continue
-
-        # If both filters don't apply, add this granule to the list
-        filtered.append(granule)
-
-    return filtered
-
-
-def filter_granules_rtc(granules, args):
-    filtered_granules = []
-    for granule in granules:
-        granule_id = granule.get("granule_id")
-
-        if COLLECTION_TO_PRODUCT_TYPE_MAP[args.collection] == "RTC":
-            match_product_id = re.match(rtc_granule_regex, granule_id)
-            burst_id = match_product_id.group("burst_id")
-
-            mgrs = mbc_client.cached_load_mgrs_burst_db(filter_land=True)
-            mgrs_sets = mbc_client.burst_id_to_mgrs_set_ids(mgrs,
-                                                            mbc_client.product_burst_id_to_mapping_burst_id(burst_id))
-            if not mgrs_sets:
-                logging.debug(f"{burst_id=} not associated with land or land/water data. skipping.")
-                continue
-
-        filtered_granules.append(granule)
-    return filtered_granules
-
-def download_from_s3(bucket, file, path):
-    s3 = boto3.resource('s3')
-    try:
-        s3.Object(bucket, file).download_file(path)
-    except Exception as e:
-        raise Exception("Exception while fetching disp frame map json file: %s. " % file + str(e))
